feat(*): Modularised existing code BREAKING CHANGE: Moved db reference path to db/ark-en.sqlite
